requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pyjwt==2.3.0
cryptography==3.4.7
pydantic==1.8.2
prometheus-client==0.12.0
python-json-logger==2.0.2
aiohttp==3.8.1
python-multipart==0.0.5
python-dotenv==0.19.0
pyyaml==5.4.1 